requirements.txt
flask
flask-cors
requests
torch
torchvision
pillow
